Updated Hero & Hero Scripts
The Hero's sprite have been finished for now, to make the hero feel more
alive rotation angles have been added to make the herolook the way the
hero is walking. Because of this change the camera had to be taken away
from the hero to keep that from rotating, but to keep the camera on the
hero a new script CameraFollow was written.