started creating combat
Made collision with enemies and combat start happen. Creating a list of
enemies and spawning them at a combat place far away from the main
level, as well as moving the player there on collision. Had to move a
bunch of stuff to make it work, that's why there are so many deletes.